+++ packages/analytics/__tests__/utils/resolveCredentials.test.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { fetchAuthSession } from '@aws-amplify/core';
+
 import { resolveCredentials } from '../../src/utils';
 import { AnalyticsError } from '../../src';