+++ packages/notifications/__tests__/pushNotifications/providers/pinpoint/apis/setBadgeCount.test.ts
@@ -6,6 +6,8 @@ import { expectNotSupported } from '../../../../testUtils/expectNotSupported';
 
 describe('getBadgeCount', () => {
 	it('is only supported on React Native', () => {
-		expectNotSupported(() => setBadgeCount(42));
+		expectNotSupported(() => {
+			setBadgeCount(42);
+		});
 	});
 });