+++ packages/storage/__tests__/providers/s3/apis/uploadData/multipartHandlers.test.ts
@@ -21,6 +21,7 @@ import { UPLOADS_STORAGE_KEY } from '../../../../../src/providers/s3/utils/const
 import { byteLength } from '../../../../../src/providers/s3/apis/uploadData/byteLength';
 import { CanceledError } from '../../../../../src/errors/CanceledError';
 import { StorageOptions } from '../../../../../src/types';
+import '../testUtils';
 
 jest.mock('@aws-amplify/core');
 jest.mock('../../../../../src/providers/s3/utils/client');
@@ -40,12 +41,12 @@ const defaultCacheKey = '8388608_application/octet-stream_bucket_public_key';
 const testPath = 'testPath/object';
 const testPathCacheKey = `8388608_${defaultContentType}_${bucket}_custom_${testPath}`;
 
-const mockCreateMultipartUpload = createMultipartUpload as jest.Mock;
-const mockUploadPart = uploadPart as jest.Mock;
-const mockCompleteMultipartUpload = completeMultipartUpload as jest.Mock;
-const mockAbortMultipartUpload = abortMultipartUpload as jest.Mock;
-const mockListParts = listParts as jest.Mock;
-const mockHeadObject = headObject as jest.Mock;
+const mockCreateMultipartUpload = jest.mocked(createMultipartUpload);
+const mockUploadPart = jest.mocked(uploadPart);
+const mockCompleteMultipartUpload = jest.mocked(completeMultipartUpload);
+const mockAbortMultipartUpload = jest.mocked(abortMultipartUpload);
+const mockListParts = jest.mocked(listParts);
+const mockHeadObject = jest.mocked(headObject);
 
 const disableAssertionFlag = true;
 
@@ -55,20 +56,23 @@ const mockMultipartUploadSuccess = (disableAssertion?: boolean) => {
 	let totalSize = 0;
 	mockCreateMultipartUpload.mockResolvedValueOnce({
 		UploadId: 'uploadId',
+		$metadata: {},
 	});
+	// @ts-expect-error Special mock to make uploadPart return input part number
 	mockUploadPart.mockImplementation(async (s3Config, input) => {
 		if (!disableAssertion) {
 			expect(input.UploadId).toEqual('uploadId');
 		}
 
 		// mock 2 invocation of onProgress callback to simulate progress
-		s3Config?.onUploadProgress({
-			transferredBytes: input.Body.byteLength / 2,
-			totalBytes: input.Body.byteLength,
+		const body = input.Body as ArrayBuffer;
+		s3Config?.onUploadProgress?.({
+			transferredBytes: body.byteLength / 2,
+			totalBytes: body.byteLength,
 		});
-		s3Config?.onUploadProgress({
-			transferredBytes: input.Body.byteLength,
-			totalBytes: input.Body.byteLength,
+		s3Config?.onUploadProgress?.({
+			transferredBytes: body.byteLength,
+			totalBytes: body.byteLength,
 		});
 
 		totalSize += byteLength(input.Body)!;
@@ -80,9 +84,11 @@ const mockMultipartUploadSuccess = (disableAssertion?: boolean) => {
 	});
 	mockCompleteMultipartUpload.mockResolvedValueOnce({
 		ETag: 'etag',
+		$metadata: {},
 	});
 	mockHeadObject.mockResolvedValueOnce({
 		ContentLength: totalSize,
+		$metadata: {},
 	});
 };
 
@@ -91,8 +97,10 @@ const mockMultipartUploadCancellation = (
 ) => {
 	mockCreateMultipartUpload.mockImplementation(async () => ({
 		UploadId: 'uploadId',
+		$metadata: {},
 	}));
 
+	// @ts-expect-error Only need partial mock
 	mockUploadPart.mockImplementation(async ({ abortSignal }, { PartNumber }) => {
 		beforeUploadPartResponseCallback?.();
 		if (abortSignal?.aborted) {
@@ -105,10 +113,13 @@ const mockMultipartUploadCancellation = (
 		};
 	});
 
-	mockAbortMultipartUpload.mockResolvedValueOnce({});
+	mockAbortMultipartUpload.mockResolvedValueOnce({
+		$metadata: {},
+	});
 	// Mock resumed upload and completed upload successfully
 	mockCompleteMultipartUpload.mockResolvedValueOnce({
 		ETag: 'etag',
+		$metadata: {},
 	});
 };
 
@@ -194,7 +205,9 @@ describe('getMultipartUploadHandlers with key', () => {
 						options: options as StorageOptions,
 					});
 					const result = await multipartUploadJob();
-					expect(mockCreateMultipartUpload).toHaveBeenCalledWith(
+					await expect(
+						mockCreateMultipartUpload,
+					).toBeLastCalledWithConfigAndInput(
 						expect.objectContaining({
 							credentials,
 							region,
@@ -258,7 +271,9 @@ describe('getMultipartUploadHandlers with key', () => {
 			expect(mockCreateMultipartUpload).toHaveBeenCalledTimes(1);
 			expect(mockUploadPart).toHaveBeenCalledTimes(10_000);
 			expect(mockCompleteMultipartUpload).toHaveBeenCalledTimes(1);
-			expect(mockUploadPart.mock.calls[0][1].Body.byteLength).toEqual(10 * MB); // The part size should be adjusted from default 5MB to 10MB.
+			expect(
+				(mockUploadPart.mock.calls[0][1].Body as ArrayBuffer).byteLength,
+			).toEqual(10 * MB); // The part size should be adjusted from default 5MB to 10MB.
 		});
 
 		it('should throw error when remote and local file sizes do not match upon completed upload', async () => {
@@ -267,6 +282,7 @@ describe('getMultipartUploadHandlers with key', () => {
 			mockHeadObject.mockReset();
 			mockHeadObject.mockResolvedValue({
 				ContentLength: 1,
+				$metadata: {},
 			});
 
 			const { multipartUploadJob } = getMultipartUploadHandlers(
@@ -318,6 +334,7 @@ describe('getMultipartUploadHandlers with key', () => {
 			mockUploadPart.mockReset();
 			mockUploadPart.mockResolvedValueOnce({
 				ETag: `etag-1`,
+				// @ts-expect-error Special mock to make uploadPart return input part number.
 				PartNumber: 1,
 			});
 			mockUploadPart.mockRejectedValueOnce(new Error('error'));
@@ -370,7 +387,7 @@ describe('getMultipartUploadHandlers with key', () => {
 				}),
 			);
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -388,7 +405,7 @@ describe('getMultipartUploadHandlers with key', () => {
 
 		it('should cache the upload with file including file lastModified property', async () => {
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -423,7 +440,7 @@ describe('getMultipartUploadHandlers with key', () => {
 				}),
 			);
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -441,7 +458,7 @@ describe('getMultipartUploadHandlers with key', () => {
 
 		it('should cache upload task if new upload task is created', async () => {
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -468,7 +485,7 @@ describe('getMultipartUploadHandlers with key', () => {
 
 		it('should remove from cache if upload task is completed', async () => {
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -490,7 +507,7 @@ describe('getMultipartUploadHandlers with key', () => {
 		it('should remove from cache if upload task is canceled', async () => {
 			expect.assertions(2);
 			mockMultipartUploadSuccess(disableAssertionFlag);
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -612,6 +629,7 @@ describe('getMultipartUploadHandlers with key', () => {
 			);
 			mockListParts.mockResolvedValue({
 				Parts: [{ PartNumber: 1 }],
+				$metadata: {},
 			});
 
 			const onProgress = jest.fn();
@@ -701,7 +719,9 @@ describe('getMultipartUploadHandlers with path', () => {
 						data: twoPartsPayload,
 					});
 					const result = await multipartUploadJob();
-					expect(mockCreateMultipartUpload).toHaveBeenCalledWith(
+					await expect(
+						mockCreateMultipartUpload,
+					).toBeLastCalledWithConfigAndInput(
 						expect.objectContaining({
 							credentials,
 							region,
@@ -765,7 +785,9 @@ describe('getMultipartUploadHandlers with path', () => {
 			expect(mockCreateMultipartUpload).toHaveBeenCalledTimes(1);
 			expect(mockUploadPart).toHaveBeenCalledTimes(10_000);
 			expect(mockCompleteMultipartUpload).toHaveBeenCalledTimes(1);
-			expect(mockUploadPart.mock.calls[0][1].Body.byteLength).toEqual(10 * MB); // The part size should be adjusted from default 5MB to 10MB.
+			expect(
+				(mockUploadPart.mock.calls[0][1].Body as ArrayBuffer).byteLength,
+			).toEqual(10 * MB); // The part size should be adjusted from default 5MB to 10MB.
 		});
 
 		it('should throw error when remote and local file sizes do not match upon completed upload', async () => {
@@ -774,6 +796,7 @@ describe('getMultipartUploadHandlers with path', () => {
 			mockHeadObject.mockReset();
 			mockHeadObject.mockResolvedValue({
 				ContentLength: 1,
+				$metadata: {},
 			});
 
 			const { multipartUploadJob } = getMultipartUploadHandlers(
@@ -825,6 +848,7 @@ describe('getMultipartUploadHandlers with path', () => {
 			mockUploadPart.mockReset();
 			mockUploadPart.mockResolvedValueOnce({
 				ETag: `etag-1`,
+				// @ts-expect-error Special mock to make uploadPart return input part number.
 				PartNumber: 1,
 			});
 			mockUploadPart.mockRejectedValueOnce(new Error('error'));
@@ -877,7 +901,7 @@ describe('getMultipartUploadHandlers with path', () => {
 				}),
 			);
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -895,7 +919,7 @@ describe('getMultipartUploadHandlers with path', () => {
 
 		it('should cache the upload with file including file lastModified property', async () => {
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -933,7 +957,7 @@ describe('getMultipartUploadHandlers with path', () => {
 				}),
 			);
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -951,7 +975,7 @@ describe('getMultipartUploadHandlers with path', () => {
 
 		it('should cache upload task if new upload task is created', async () => {
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -976,7 +1000,7 @@ describe('getMultipartUploadHandlers with path', () => {
 
 		it('should remove from cache if upload task is completed', async () => {
 			mockMultipartUploadSuccess();
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -998,7 +1022,7 @@ describe('getMultipartUploadHandlers with path', () => {
 		it('should remove from cache if upload task is canceled', async () => {
 			expect.assertions(2);
 			mockMultipartUploadSuccess(disableAssertionFlag);
-			mockListParts.mockResolvedValueOnce({ Parts: [] });
+			mockListParts.mockResolvedValueOnce({ Parts: [], $metadata: {} });
 			const size = 8 * MB;
 			const { multipartUploadJob } = getMultipartUploadHandlers(
 				{
@@ -1120,6 +1144,7 @@ describe('getMultipartUploadHandlers with path', () => {
 			);
 			mockListParts.mockResolvedValue({
 				Parts: [{ PartNumber: 1 }],
+				$metadata: {},
 			});
 
 			const onProgress = jest.fn();