+++ packages/storage/__tests__/providers/s3/utils/client/S3/getPresignedGetObjectUrl.test.ts
@@ -5,7 +5,7 @@ import { presignUrl } from '@aws-amplify/core/internals/aws-client-utils';
 
 import { getPresignedGetObjectUrl } from '../../../../../../src/providers/s3/utils/client';
 
-import { defaultConfig } from './cases/shared';
+import { defaultConfigWithStaticCredentials } from './cases/shared';
 
 jest.mock('@aws-amplify/core/internals/aws-client-utils', () => {
 	const original = jest.requireActual(
@@ -25,8 +25,8 @@ describe('serializeGetObjectRequest', () => {
 	it('should return get object API request', async () => {
 		const actual = await getPresignedGetObjectUrl(
 			{
-				...defaultConfig,
-				signingRegion: defaultConfig.region,
+				...defaultConfigWithStaticCredentials,
+				signingRegion: defaultConfigWithStaticCredentials.region,
 				signingService: 's3',
 				expiration: 900,
 				userAgentValue: 'UA',
@@ -38,7 +38,7 @@ describe('serializeGetObjectRequest', () => {
 		);
 		const actualUrl = actual;
 		expect(actualUrl.hostname).toEqual(
-			`bucket.s3.${defaultConfig.region}.amazonaws.com`,
+			`bucket.s3.${defaultConfigWithStaticCredentials.region}.amazonaws.com`,
 		);
 		expect(actualUrl.pathname).toEqual('/key');
 		expect(actualUrl.searchParams.get('X-Amz-Expires')).toEqual('900');
@@ -51,8 +51,8 @@ describe('serializeGetObjectRequest', () => {
 	it('should call presignUrl with uriEscapePath param set to false', async () => {
 		await getPresignedGetObjectUrl(
 			{
-				...defaultConfig,
-				signingRegion: defaultConfig.region,
+				...defaultConfigWithStaticCredentials,
+				signingRegion: defaultConfigWithStaticCredentials.region,
 				signingService: 's3',
 				expiration: 900,
 				userAgentValue: 'UA',