+++ packages/api-rest/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.35](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.34...@aws-amplify/api-rest@4.0.35) (2024-06-07)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.34](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.33...@aws-amplify/api-rest@4.0.34) (2024-06-04)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.33](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.32...@aws-amplify/api-rest@4.0.33) (2024-05-23)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.32](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.31...@aws-amplify/api-rest@4.0.32) (2024-05-23)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## [4.0.31](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.30...@aws-amplify/api-rest@4.0.31) (2024-05-13)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
 ## [4.0.30](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@4.0.29...@aws-amplify/api-rest@4.0.30) (2024-05-07)
 
 **Note:** Version bump only for package @aws-amplify/api-rest