+++ packages/notifications/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.35](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.34...@aws-amplify/notifications@2.0.35) (2024-06-07)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
+## [2.0.34](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.33...@aws-amplify/notifications@2.0.34) (2024-06-04)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
+## [2.0.33](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.32...@aws-amplify/notifications@2.0.33) (2024-05-23)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
+## [2.0.32](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.31...@aws-amplify/notifications@2.0.32) (2024-05-23)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
+## [2.0.31](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.30...@aws-amplify/notifications@2.0.31) (2024-05-13)
+
+### Bug Fixes
+
+- Wait for endpoint creation to identify user ([#13353](https://github.com/aws-amplify/amplify-js/issues/13353)) ([f17cdf0](https://github.com/aws-amplify/amplify-js/commit/f17cdf0077387ee275bea245632f1a9b88089e1e))
+
 ## [2.0.30](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.29...@aws-amplify/notifications@2.0.30) (2024-05-07)
 
 ### Bug Fixes