+++ packages/storage/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.4.5...@aws-amplify/storage@6.4.6) (2024-06-07)
+
+**Note:** Version bump only for package @aws-amplify/storage
+
+## [6.4.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.4.4...@aws-amplify/storage@6.4.5) (2024-06-04)
+
+### Bug Fixes
+
+- **storage:** Fix MD5 calculation ([#13458](https://github.com/aws-amplify/amplify-js/issues/13458)) ([30b5b85](https://github.com/aws-amplify/amplify-js/commit/30b5b851f3562a338d6afb82a3a98835ac78ccfb))
+
+## [6.4.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.4.3...@aws-amplify/storage@6.4.4) (2024-05-23)
+
+**Note:** Version bump only for package @aws-amplify/storage
+
+## [6.4.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.4.2...@aws-amplify/storage@6.4.3) (2024-05-23)
+
+### Bug Fixes
+
+- **storage:** credentials expires after 1 hour ([#13329](https://github.com/aws-amplify/amplify-js/issues/13329)) ([d7d33da](https://github.com/aws-amplify/amplify-js/commit/d7d33da192ff610c01fcdd417460ddd79d3dcb8d))
+
+## [6.4.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.4.1...@aws-amplify/storage@6.4.2) (2024-05-13)
+
+**Note:** Version bump only for package @aws-amplify/storage
+
 ## [6.4.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.4.0...@aws-amplify/storage@6.4.1) (2024-05-07)
 
 **Note:** Version bump only for package @aws-amplify/storage